+++ PadDown/tests/test_decrypt_engine.py
@@ -1,28 +1,31 @@
 import pytest
 from Crypto.Util.Padding import unpad
 from PadDown.decrypt_engine import DecryptEngine
+from PadDown.exceptions import PadDownException
 
 from ..examples.vulnerable_encryption_service import InvalidPadding, VulnerableEncryptionService
 
 VEC = VulnerableEncryptionService()
 
 
-class TestVulnerableEncryptionService:
-    def test_encryption_and_decryption(self):
-        plaintext_misaligned = b"Misaligned plaintext!"
-        ciphertext = VEC.encrypt(plaintext_misaligned)
-        answer = VEC.decrypt(ciphertext)
-        assert answer == "Decryption successful!"
-
-
 class TestDecryptEngine:
-    def test_decrypt_at_index(self):
+    def test_find_c_prime_at_index(self):
         class MyDecryptEngine(DecryptEngine):
             def has_valid_padding(self, ciphertext):
                 return ciphertext == b"\x05"
 
         decrypt_engine = MyDecryptEngine(b"dummy")
         decrypt_engine.find_c_prime_at_index(bytearray(b"\x00"), 0)
+
+    def test_exception_raised_on_bad_implementation(self):
+        class MyDecryptEngine(DecryptEngine):
+            def has_valid_padding(self, ciphertext):
+                # No valid encryption ever found
+                return False
+
+        decrypt_engine = MyDecryptEngine(b"dummy")
+        with pytest.raises(PadDownException):
+            decrypt_engine.find_c_prime_at_index(bytearray(b"\x00"), 0)
 
     @pytest.mark.skip
     def test_decrypt_block(self):
